alert user

+++ src/app/app.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { AuthenService } from './services/authen.service';
 import { Router } from '@angular/router';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-root',
@@ -9,17 +10,23 @@ import { Router } from '@angular/router';
 })
 export class AppComponent {
   title = 'nguon-api';
-  constructor(private authen: AuthenService, private route: Router){ }
+  constructor(private authen: AuthenService,
+              private route: Router,
+              private snakbar: MatSnackBar){}
   isLogin(): boolean{
     return localStorage.getItem('data')? true : false;
   }
 
   logOut(){
-    this.authen.logout().subscribe( data => {
-      console.log(data);
-    });
-    localStorage.clear();
-    this.route.navigate(['/login']);
+    if(localStorage.getItem('data')){
+      this.authen.logout().subscribe( data => {
+        localStorage.clear();
+        this.route.navigate(['/login']);
+        this.snakbar.open('Logged out!','OK',{duration: 2000});
+      });
+    }else{
+      this.snakbar.open('Unauthorized for Logout!','OK',{duration: 2000});
+    }
   }
 
 }
